requirements.txt
aiogram==3.17.0
requests==2.32.3
python-decouple==3.8